Tilføjet og virker: - Alle http metoder til Riot API er nu samlet i ExternalAPIManagers.package > RiotAPIManager.java
